+++ CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG for Sulu
     * BUGFIX      #2697 [ContentBundle]       Deindex page after unpublishing
     * BUGFIX      #2684 [ContentBundle]       Disabled options in toolbar item which are not avialable when editing a page
     * FEATURE     #2689 [ContentBundle]       Added functionality to delete a draft
+    * BUGFIX      #2683 [MediaBundle]         Removed selected datagrid from media selection overlay
     * FEATURE     #2559 [CoreBundle]          Renamed parameters.yml to parameters.yml.dist so you can use a local version
     * BUGFIX      #2678 [ContentBundle]       Fixed error caused by draft label when opening a ghost page
     * BUGFIX      #2668 [ContentBundle]       Fixed resource locator generation for pages with unpublished parents